Remove ordered map Codec.
Thinking more about it, this was something that shouldn't have been
done. I just can't imagine a use case for defining an ordering between
maps where a struct or slice wouldn't be better.

This will also allow us to remove the key-value pair reader/writer
types, and use TerminateIfNeeded() in the map Codec.

Ref #28