Fix a bug in the negate codec, and add tests to catch a regression.
Negation needs to not only reverse the ordering of bytes, it also
needs to reverse the ordering of a []byte and its prefixes. Normally,
"ab" < "abc". When negated, it needs to be the opposite of that,
simply flipping bits is not enough.

The bug was in the order of operations, escape then flip instead of
flip then escape.