Add some more zero-length array tests.
This was in preparation for possibly removing the empty prefix, and
having only nil/non-nil. But zero-length arrays do not have an empty
encoding because the array codec delegates to the pointer-to-array
codec, which always writes a nil/non-empty prefix.

If this works out everywhere, the non-empty prefix would become
non-nil instead.

Ref #31