New baseline.
There is a bit or weirdness where the same exact code performed up to
around 10% worse. I have no idea why yet. The benchmarks were run in a
terminal with nothing else running outside whatever the OS was doing.
Next time I take a baseline, I'm going to take the machine offline to
see if that reduces the variance.

Some of the outliers:
- BigInt: get() is worse
- BigFloat: all worse except big_neg
  however, there is still a fairly big optimization to be made here
- Terminate: worse for smaller values
  vs. some intermediate results, same code

These are worse, with the same exact code.
- SliceOf, CastSliceOf: 1k/append_nil,get
- MapOf: empty/get; 1/get; 1k/append_nil,put,get
- PointerTo: *0/put/get, *-1/put/get